allow more than one negotiator #5967 a new knob. COLLECTOR_ALLOW_ONLY_ONE_NEGOTIATOR will enable the old behavior. it defaults to false. this commit also changes the default name of the negotiator from $(HOSTNAME) to $(SUBSYS) or $(LOCALNAME) and allows it to be set via the new NEGOTIATOR_NAME knob. ===VersionHistory:Pending===

+++ src/condor_collector.V6/collector_engine.cpp
@@ -82,6 +82,8 @@ CollectorEngine::CollectorEngine (CollectorStats *stats ) :
 	m_forwardFilteringEnabled = false;
 	housekeeperTimerID = -1;
 
+	m_allowOnlyOneNegotiator = param_boolean("COLLECTOR_ALLOW_ONLY_ONE_NEGOTIATOR", false);
+
 	collectorStats = stats;
 	m_collector_requirements = NULL;
 }
@@ -759,10 +761,12 @@ collect (int command,ClassAd *clientAd,const condor_sockaddr& from,int &insert,S
 			break;
 		}
 		hashString.Build( hk );
+		if (m_allowOnlyOneNegotiator) {
 			// first, purge all the existing negotiator ads, since we
 			// want to enforce that *ONLY* 1 negotiator is in the
 			// collector any given time.
-		purgeHashTable( NegotiatorAds );
+			purgeHashTable( NegotiatorAds );
+		}
 		retVal=updateClassAd (NegotiatorAds, "NegotiatorAd  ", "Negotiator",
 							  clientAd, hk, hashString, insert, from );
 		break;
